[Test] Split long runtime env tests.
