+++ python/ray/tests/BUILD
@@ -133,6 +133,7 @@ py_test_module_list(
   files = [
     "test_placement_group.py",
     "test_runtime_env.py",
+    "test_runtime_env_2.py"
   ],
   size = "large",
   extra_srcs = SRCS,
